Cleaning accounts cache when switching networks.
