+++ background/transport.js
@@ -60,9 +60,22 @@ let accountsCache = {};
 function newTransport (token) {
   const transport = new Ws(`ws://${UI}`, token, true);
   transport.on('open', () => {
+    let oldOrigins = Object.keys(accountsCache);
     accountsCache = {};
+    // re-populate cache (for new network)
+    oldOrigins.forEach(origin => {
+      fetchAccountsForCache(origin);
+    });
   });
   return transport;
+}
+
+function fetchAccountsForCache (origin) {
+  return transport.execute('parity_getDappsAddresses', origin)
+    .then(accounts => {
+      accountsCache[origin] = accounts;
+      return accounts;
+    });
 }
 
 chrome.runtime.onMessage.addListener((request, sender, callback) => {
@@ -83,9 +96,8 @@ chrome.runtime.onMessage.addListener((request, sender, callback) => {
       });
     }
 
-    transport.execute('parity_getDappsAddresses', origin)
+    fetchAccountsForCache(origin)
       .then(accounts => {
-        accountsCache[origin] = accounts;
         return callback({
           err: null,
           payload: accounts
@@ -103,7 +115,7 @@ chrome.runtime.onMessage.addListener((request, sender, callback) => {
 
   if (!isTransportReady && request.token) {
     if (transport) {
-      // TODO [ToDr] kill old transport!
+      transport.close();
     }
 
     if (openedTabId) {
@@ -144,6 +156,9 @@ function extractToken () {
       return;
     }
 
+    if (transport) {
+      transport.close();
+    }
     transport = newTransport(token.authToken);
   });
 }
