+++ manifest.json
@@ -7,7 +7,7 @@
   "content_scripts": [
     {
       "matches": ["<all_urls>"],
-      "js": ["web3/index.js", "web3/inpage.js"],
+      "js": ["web3/content_script.js", "web3/index.js", "web3/inpage.js"],
       "run_at": "document_start"
     },
     {
